✨ new LatestSales component
