+++ utils/icons.ts
@@ -68,7 +68,6 @@ import {
   faQrcode,
   faFire,
   faMoneyBill,
-  faMoneyBillWave,
   faHistory,
   faEllipsisH,
   faLightbulb,
@@ -166,7 +165,6 @@ library.add(
   faTh,
   faFire,
   faMoneyBill,
-  faMoneyBillWave,
   faHistory,
   faEllipsisH,
   faLightbulb,
